+++ src/OmnichannelChatSDK.ts
@@ -1727,6 +1727,10 @@ class OmnichannelChatSDK {
             requestOptionalParams.initContext!.preChatResponse = optionalParams.preChatResponse;
         }
 
+        if (optionalParams.portalContactId) {
+            requestOptionalParams.initContext!.portalcontactid = optionalParams.portalContactId;
+        }
+
         if (optionalParams.sendDefaultInitContext) {
             if (platform.isNode() || platform.isReactNative()) {
                 const exceptionDetails: ChatSDKExceptionDetails = {